refactor: move ActivityRepository
